Created a new branch for EFM 4.2 and copied the 4.1 files to it
We need to work on the upcoming product releases, for that we need a next version of our product docs. I tried creating a new folder 4.2 (upcoming release) parrallel to 4.1 (latest release) and copied the content from 4.1 branch. If this works, then I can start updating the content for EFM 4.2 branch.


Former-commit-id: 54aba4ef9e0cfa9f225c238ca1e9fa09c8b156c5